Dimitris/add circuit for benchmarking (#5)
* Implemented circuit for benchmarking

* Cleanup

* Moved circuit to binary target

* Configured binary to accept power as command line argument

* Fixed comments

* Fixed comments

* Updated code

* Rename circuit

* Updated fibonacci_bench.rs

* Debugged code

* Handled case where compress flag is set

* Fixed bug involving the 32-bit runtime

* Updated benchmarking binary and script

* Bumped up version to 0.2.1

* Updated description

+++ src/deserializer/mod.rs
@@ -63,7 +63,7 @@ where
     F: RichField + Extendable<D>,
     C: GenericConfig<D, F = F>,
 {
-    let offset = size_of::<usize>();
+    let offset = size_of::<u64>();
     if pubs.len() < offset {
         return Err(DeserializeError::InvalidProof);
     }
